fix: fix mock get_entry signature
Fixes failing unit test test_transfer_usd_near.
The root cause was get_entry method signature out of sync in conversion
proxy contract and in mock fpo contract.

It can be easily detected by and reported by removing the `.then` part
in the conversion_proxy/src/lib.rs#L107-116 . After doing that you can
use println in the unit test to print the outcome and see the reason of
the fail for the original root cause.